Fix regression test for amd64
Changes:
* Several places in our code assume that int and long have the same size. This is not true for amd64 and causes regtest failures.
* The acc_flush code used to do buggy errno handling which resulted in ACHAT_ERROR instead of ACHAT_PENDING in some cases. Fix
this and addapt the test.
* The error code translation test used hard coded names for LC_MESSAGES that are wrong if the target system is utf8 (true for
karmic-amd64 chroot).
* The helper utilities in ubuntu-amd64/utils/bin/ were not built by default.

Aegis-Change: anoubis.10.972.C587 Delta 295
Bug: 1481
BLE

+++ test/libapn/test_rbtree.c
@@ -87,9 +87,9 @@ int main()
 	for (try = 0; cnt || try < 1000000; ++try) {
 		int idx = rand() % NENT;
 		if (try > 1000000)
-			for (; e[idx].key == (unsigned int)-1;
+			for (; e[idx].key == (unsigned long)-1;
 			    idx = (idx+1) % NENT);
-		if (e[idx].key == (unsigned int)-1) {
+		if (e[idx].key == (unsigned long)-1) {
 			e[idx].key = idx;
 			assert(rb_insert_entry(&root, e+idx));
 			cnt++;
